+++ srcs/fdf/projection/projection_calcule.c
@@ -6,14 +6,14 @@
 /*   By: obouhlel <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/03 10:02:09 by obouhlel          #+#    #+#             */
-/*   Updated: 2023/01/07 13:21:29 by obouhlel         ###   ########.fr       */
+/*   Updated: 2023/01/07 18:07:26 by obouhlel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../../includes/fdf.h"
 
-t_projection	*ft_calcule_projection(t_map *map, double **mp, \
-										double x, double y)
+t_projection	*ft_calcule_projection(t_map *map, float **mp, \
+										float x, float y)
 {
 	t_projection	*projection;
 	t_projection	*new;
